move get result and logs to neontxresinfo
